Improvements to analyze_language and main:
No longer attempting to do grammar analysis.

We now have data on most common words for the five major parts of speech.

We no longer have redundancy in get_sentiment(); however, it's still taking a ridiculous amount of time and I will need to work on reducing that, deciding whether it's worth the wait, or taking part of it out.

TBD:

Restructuring main/analyze_language so that the data is collected more neatly--it's a little sloppy and confusing right now.

Completing the unittest for analyze_language.