So my stuff is now mostly done!
I've reorganized main.py so that the data is not collected in a horribly ugly way. I've made a few fixes to analyze_language (we were having issues with *, [, and ] etc being tagged as nouns, which obviously isn't helpful; that's been rectified).

I am now done with my section of main, but if I have free time I will come back and implement help/debug.

I am mostly done with analyze_language as well, but my unit test is not complete, and I still have not figured out how to make sentiment analysis run in a reasonable amount of time.